backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
aiofiles==23.2.1
numpy==1.26.4
faiss-cpu==1.13.2

# Local LLM and embeddings
httpx>=0.25.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Gemini API (for cloud LLM option)
google-genai>=1.0.0
